requirements.txt
Django==5.2.8
pymongo==4.6.0
python-dotenv==1.0.0